refactor(search): introduce variant type w/pattern matching on filters
Primarily to take more advantage of the language by getting more help
from the compiler to catch missing functionality due to exhaustive
pattern matching.

The new approach will do a better job in enforcing correctness when
new search filters are added later. Compared to the old approach which
the compiler couldn't help us much because it didn't know about all the
filters that we were supposed to support.