refactor(build): introduce @zeit/ncc to prebuild .re to .js
These changes does a second try of prebuilding all ReasonML source files,
including all dependency, to a single JavaScript file. This is done to
prevent end-users having to download the entire Bucklescript platform
and build the application on their computer, but rather download a single
file that has already been built and is ready for action.

Closes https://github.com/phillipj/gitlab-search/issues/4