refactor(ci): kick off build process in npm postinstall -> prepublish
Because we always want to ensure that all files has been built before
publishing this CLI to npm, we move the automatic build process from
`$ npm install` to just-before publishing a new version to npmjs.

This also relates to just having re-introduced prebuilding from .re to
.js, where end-users no longer have to download all 3rd party deps and
build the application on their own computer.