Added page limit to the GitLab API calls
GitLab APIs for group and project lists default to 20 items per page. The ideal solution would be to implement paging to the API calls. As a short term fix that should cover the majority of scenarios, setting the items per page to 1000 is a quick and easy change that should resolve the issue in the majority of cases.