feat: add --concurrency option for custom requests limit (default: `25`)
Recently changes has been introduced to limit the amount of concurrent
HTTPS requests being sent to GitLab.

That was proven vital after fixing the limitation where only the first
100 project were search in, as sending several hundreds concurrent
requests to GitLab often resulted in 502 responses, in practise DDoSing
GitLab when performing search queries.

The default is set to `25` for now. That's a shot in the dark basically,
as we've got no idea how many concurrent requests GitLab are supposed to
handle. Regardless, having on-premise installations of GitLab in mind,
where the amount of resources would vary a lot, there's really no
silverbullet here.

Hopefully it'll work as inteded for most people, meaning not too many
would have to think about adjusting this concurrency limit at all.