Rewrite from ReasonML to TypeScript
These changes are inspired by the fact that ReasonML was a very
interesting learning experience years back and was a cool challenge,
but as time as passed, the ReasonML aspect of this tool has severily
held back external contributions.

Since the initial creator has stopped using GitLab the recent years,
it's even more important for this project to be approachable by new
collaborators, where TypeScript certainly opens up the possibilities
a lot more.

Lots of help from our new AI companions, the .ts files are meant to be
a direct translation of the old .re variants. Few nitpicky changes has
been done after initial AI translation, let's see if it works in
practice or not.